+++ configure.py
@@ -938,6 +938,7 @@ scylla_core = (['message/messaging_service.cc',
                 'db/data_listeners.cc',
                 'db/functions/function.cc',
                 'db/hints/internal/hint_endpoint_manager.cc',
+                'db/hints/internal/hint_sender.cc',
                 'db/hints/manager.cc',
                 'db/hints/resource_manager.cc',
                 'db/hints/host_filter.cc',
