+++ db/CMakeLists.txt
@@ -16,6 +16,7 @@ target_sources(db
     data_listeners.cc
     functions/function.cc
     hints/internal/hint_endpoint_manager.cc
+    hints/internal/hint_sender.cc
     hints/manager.cc
     hints/resource_manager.cc
     hints/host_filter.cc
